+++ src/redturtle/prenotazioni/tests/test_delete_booking_api.py
@@ -116,7 +116,7 @@ class TestDeleteBookingApi(unittest.TestCase):
         self.assertEqual(404, response.status_code)
 
     def test_admin_can_delete_booking(self):
-        booking = self.booker.create(
+        booking = self.booker.book(
             {
                 "booking_date": self.today + timedelta(1),  # tomorrow
                 "booking_type": "Type A",
@@ -136,7 +136,7 @@ class TestDeleteBookingApi(unittest.TestCase):
 
     @unittest.skip("wip")
     def test_anon_cant_delete_other_user_booking(self):
-        booking = self.booker.create(
+        booking = self.booker.book(
             {
                 "booking_date": self.today + timedelta(1),  # tomorrow
                 "booking_type": "Type A",
@@ -155,7 +155,7 @@ class TestDeleteBookingApi(unittest.TestCase):
 
     def test_anon_can_delete_anon_booking(self):
         logout()
-        booking = self.booker.create(
+        booking = self.booker.book(
             {
                 "booking_date": self.today + timedelta(1),  # tomorrow
                 "booking_type": "Type A",
@@ -180,7 +180,7 @@ class TestDeleteBookingApi(unittest.TestCase):
 
     def test_user_can_delete_his_booking(self):
         login(self.portal, "user")
-        booking = self.booker.create(
+        booking = self.booker.book(
             {
                 "booking_date": self.today + timedelta(1),  # tomorrow
                 "booking_type": "Type A",
@@ -206,7 +206,7 @@ class TestDeleteBookingApi(unittest.TestCase):
     @unittest.skip("wip")
     def test_user_cant_delete_other_user_booking(self):
         login(self.portal, "user")
-        booking = self.booker.create(
+        booking = self.booker.book(
             {
                 "booking_date": self.today + timedelta(1),  # tomorrow
                 "booking_type": "Type A",
@@ -232,7 +232,7 @@ class TestDeleteBookingApi(unittest.TestCase):
     @unittest.skip("wip")
     def test_user_cant_delete_anon_booking(self):
         logout()
-        booking = self.booker.create(
+        booking = self.booker.book(
             {
                 "booking_date": self.today + timedelta(1),  # tomorrow
                 "booking_type": "Type A",
@@ -256,7 +256,7 @@ class TestDeleteBookingApi(unittest.TestCase):
         )
 
     def test_cant_delete_past_booking(self):
-        booking = self.booker.create(
+        booking = self.booker.book(
             {
                 "booking_date": self.today - timedelta(1),  # yesterday
                 "booking_type": "Type A",
@@ -283,7 +283,7 @@ class TestDeleteBookingApi(unittest.TestCase):
         )
 
     def test_cant_delete_today_booking(self):
-        booking = self.booker.create(
+        booking = self.booker.book(
             {
                 "booking_date": self.today,
                 "booking_type": "Type A",